action: Remove *BaseAction.LogStart() function
The LogStart() function was manually called at the start of each Run()
function for each action to print the initial log line for each action (to
print the action name/description). Move this logic to the main loop so
that we don't need to reply on an action to remember to call this function.

Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>