actions: run: Check script exists during recipe verification
If a script doesn't exist or the permissions are wrong, the recipe exits
during the recipe execution. Let's do some early checks in the Verify
stage to ensure the script exists on the host filesystem and a basic check
that the file has executable permission bit set. This check also follows
symlinks.

Signed-off-by: Christopher Obbard <chris.obbard@oss.qualcomm.com>