commands: set qemu path based on host architecture
Currently, we assume `debos` will run only on `amd64` hosts. However,
`fakemachine` recently gained the ability to run on `arm64` hosts, so we
shouldn't make this assumption anymore.

This change only sets the path to `qemu` when building for a foreign
arch, instead of setting it for all architectures except x86 ones.

Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.com>