debootstrap_action: don't assume the host is amd64
This change marks the target architecture as "foreign" if it differs
from the current host architecture, instead of hardcoding an expected
host architecture.

Due to Debian architectures not translating directly to their Go
equivalents (like `mips*el` being `mips*le` instead, or `i386` being
simply `386` in Go), this will only work for `amd64`, `arm64` and
`riscv64` hosts. However, it seems unlikely running `debos` on a 32-bit
host (or even on `mips64el`) will become a compelling use-case in the
future, making this issue an acceptable limitation.

Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.com>