commands: Set nspawn console mode to pipe
Currently the default nspawn console mode for nspawn is interactive which
means that /dev/console is created inside the nspawn container and the
command can assume it is running in an interactive session.

With console mode set to pipe, nspawn doesn't create /dev/console inside
the container and scripts assume they are running in an interactive
environment

This has the benefit of forcing some commands (e.g. apt) to not show things
like progress bars which only work in interactive terminals.

Fixes: #411
Fixes: #416
Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>