docker: Remove Arch support from Debian-based container
It turns out that trying to bootstrap an Arch system from Debian-packaged
tools is quite hard since Arch moves much quicker than Debian Bookworm.
Any upstream Arch churn can cause the Debos container build to break
(this has happened twice now).

Since this really is an anti-pattern in Debos - we suggest building an OS
image from the packages which make that OS - remove the Arch bits from
the Debian Dockerfile to stop breakages. The hope is as time goes on we
can create a separate Arch-based Dockerfile for Arch support.

Link: https://github.com/go-debos/debos/issues/483
Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>