ci: Add formatting linter stage to workflow
Ensure that submitted merge requests are run through a go linter check to keep
high code quality. Run the linter job in parallel with the unit test jobs
but do not run subsequent Docker container tests (which are heavy on
resources) if the linter job fails.

For now, disable all linters and simply run a `go fmt` check due to
older codebase. The intention is to enable linters one-by-one once the code
quality is manually cleaned up.

Signed-off-by: Christopher Obbard <christopher.obbard@linaro.org>